Tests/LibWeb/Text/expected/FileAPI/filereader-basic.txt

1: readAsText(): 'This is some data to be read! 🦬' error: 'null'
2: readAsDataURL(): 'data:application/octet-stream;base64,VGhpcyBpcyBzb21lIGRhdGEgdG8gYmUgcmVhZCEg8J+mrA==' error: 'null'
3: readAsArrayBuffer(): 'This is some data to be read! 🦬' error: 'null'
